% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUrl.R
\name{buildUrl}
\alias{buildUrl}
\title{Function to build url for getting reports from Infinity API}
\usage{
buildUrl(param.list)
}
\arguments{
\item{param.list}{List of all request parameters. See \code{\link{getReport}} for full example.}
}
\value{
URL for getting reports.
}
\description{
Function to build url for getting reports from Infinity API and to perform validation of some fields.
URL can be used for getting raw data in browser.
}

