% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyInfinity.R
\name{ClassifyInfinity}
\alias{ClassifyInfinity}
\title{Infinity pipeline}
\usage{
ClassifyInfinity(
  inputFile,
  outputFile = "results.csv",
  model,
  QC_value = 0.6,
  Length_value = 0.2,
  N_value = 2
)
}
\arguments{
\item{inputFile}{string path relative to the working directory of the input file. Must be in fasta format.}

\item{outputFile}{string path relative to the working directory of the output file. Default = "results.csv"}

\item{model}{A random forest classification model. Must be FULL_HA or HA1. See \link{FULL_HA} and \link{HA1}.}

\item{QC_value}{numeric value from 0 to 1. (default = 0.6)}

\item{Length_value}{numeric value from 0 to 0.4. Proportion of difference to the expected sequence length. (default = 0.2)}

\item{N_value}{numeric value from 0 to 100. Percentage of acceptable ambiguous bases. (default = 2)}
}
\value{
a comma-delimited file
}
\description{
Wrapper function that performs all the recommended step. This includes
\describe{
\item{Kcounter}{# Count Kmers and normalize data}
\item{PredictionCaller}{Run prediction pipeline}
\item{QualityControl}{Check Quality of the input data and classification results}
\item{Quality_filter}{For sequences with any FLAG present set result Clade to LowQuality}
}
}
\examples{
file_path<-system.file("extdata","test_dataset.fasta",package="infinity")
ClassifyInfinity(inputFile=file_path,model=FULL_HA)
}
