% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CounterNormalizer.R
\name{CounterNormalizer}
\alias{CounterNormalizer}
\title{CounterNormalizer}
\usage{
CounterNormalizer(SequenceData, model)
}
\arguments{
\item{model}{}
}
\value{
A list of 3 vectors: normalized k-er counts, genome length and contents of undefined bases.
}
\description{
Counts k-mers of the size required by the input model and normalize the data regarding genome size.
}
\examples{
load test dataset
file_path<-system.file("extdata","test_dataset.fasta",package="infinity")
SequenceData<-ape::read.FASTA(file_path,type = "DNA")

NormalizedData<-CounterNormalizer(SequenceData,
}
