% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIGRP.R
\name{getIGRP}
\alias{getIGRP}
\title{Get list of igrps}
\usage{
getIGRP(username, password)
}
\arguments{
\item{username}{Username to access data}

\item{password}{Password to access data}
}
\value{
Data frame which contains all installation groups (igrp), available for given user
}
\description{
Get list of igrps
}
\examples{
\dontrun{
df <- getIGRP(username, password)
}

}

