% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReport.R
\name{getReport}
\alias{getReport}
\title{Function for getting report from Infinity API}
\usage{
getReport(url, username, password, pagination = FALSE)
}
\arguments{
\item{url}{URL for getting data. See \code{\link{buildUrl}}}

\item{username}{Username}

\item{password}{Password}

\item{pagination}{Results from the Infinity API are restricted to a maximum of number of rows per request.
This maximum number is typically 10,000 and if you specify a higher limit value the API will return an error.
For getting of all records parameter pagination = TRUE.
If the number of rows in result data frame will be less than 10,000 than pagination doesn't used.}
}
\description{
Function for getting report from Infinity API
}
\examples{
\dontrun{
# Create list of parameters
params <- triggerInit(igrp = 840,
                     trigger = "calls",
                     limit = 10500,
                     sort = "triggerDatetime-desc"
                     display = c("triggerDatetime",
                                 "srcPhoneNumber",
                                 "callState",
                                 "txr",
                                 "destPhoneNumber",
                                 "ringTime"))
# Build URL for getting report
 url <- buildUrl(params)

# Getting report from Infinity API to data frame
 df <- getReport(url, username, password, pagination = TRUE)
}

}

