% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run INFINITy shiny app}
\usage{
runShinyApp()
}
\value{
a shiny app
}
\description{
Deploys a server that runs the INFINITy app locally
}
\examples{
\dontrun{
runShinyApp()
}

}
