% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCaller.R
\name{Stringent_filter}
\alias{Stringent_filter}
\title{Stringent_filter}
\usage{
Stringent_filter(data, QC_unknown = 0.2)
}
\arguments{
\item{data}{data.frame obtained with \code{\link[=PredictionCaller]{PredictionCaller()}}}

\item{QC_unknown}{numeric value from 0 to 1. It represents the minimum supported probability. If the value is not passed the Clade result is overwritten to unknown (default = 0.2)}
}
\value{
data.frame
}
\description{
Change Clade definition to unknown in samples with a probability score
below a threshold
}
\examples{

file_path<-system.file("extdata","test_dataset.fasta",package="infinity")

sequence<-ape::read.FASTA(file_path,type = "DNA")

NormalizedData <- Kcounter(SequenceData=sequence,model=FULL_HA)

PredictedData <- PredictionCaller(NormalizedData=NormalizedData,model=FULL_HA)

PredictedData <- QualityControl(PredictedData,model=FULL_HA)

PredictedData <- Quality_filter(PredictedData)
}
