% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggerInit.R
\name{triggerInit}
\alias{triggerInit}
\title{Collect request parameters}
\usage{
triggerInit(igrp = NULL, trigger = NULL, limit = NULL, offset = NULL,
  sort = NULL, filter = NULL, startDate = NULL, endDate = NULL,
  tz = NULL, display = NULL, attributionModelId = NULL,
  nestAttributions = NULL)
}
\arguments{
\item{igrp}{(int) Installation ID}

\item{trigger}{- (string) A trigger represents a single tracked event in Infinity.

Possible values:

    all  -   Get all triggers

    calls   - Get all call triggers ( where trigger action is call )

    calls/users    - Get all Hosted PBX Call User triggers ( where trigger action is callUser )

    callStatus    - Get all call status triggers ( where trigger action is callBlegStart or callAlegStop )

    goals   -  Get all goal triggers ( where trigger is marked as a goal )

    transactions   -  Get all transaction triggers ( where trigger has transaction currency or goal currency fields )#'
                     callRate triggers are excluded from these reports

    callsGoalsTransactions  -   Get all triggers with any transaction, call or goal details.
                               This is a union of the three routes "calls", "goals" and "transactions" described above

    touchPoints  -   Get all touch point triggers.
                    This is a collection of recent land, call and goal triggers.
                    You can only query back as far as your Installation's trigger retention period on this route}

\item{limit}{(int) Limit result set.}

\item{offset}{(int) Offset result set.}

\item{sort}{(array) One or more sorts. String or vector of strings.}

\item{filter}{(array) Optional. One or more filters. String or vector of strings.}

\item{startDate}{(string) Optional. Start Date to filter by, Start Date must be of the format YYYY-MM-DD.}

\item{endDate}{(string) Optional. End Date to filter by, End Date must be of the format YYYY-MM-DD.}

\item{tz}{(string) Optional. Timezone to convert to.}

\item{display}{(array) Optional. One ore more fields to display in report. String or vector of strings.}

\item{attributionModelId}{(int) Optional. Attribution Model ID to use for report. Can be 'none' to remove attribution data from report.}

\item{nestAttributions}{(int) Optional. Enables/Disables nested multiple attribution output. If disabled will return multiple rows per trigger for each attributed land.}
}
\value{
List of parameters required for getting reports
}
\description{
Function collects all request parameters in one list for passing to function \code{\link{buildUrl}}
}
\examples{
\dontrun{
params <- triggerInit(igrp = 840,
                     trigger = "calls",
                     limit = 50,
                     sort = "triggerDatetime-desc",
                     display = c("triggerDatetime",
                                 "srcPhoneNumber",
                                 "callState",
                                 "txr",
                                 "destPhoneNumber",
                                 "ringTime"))
}
}
\seealso{
Full description of filter and sort fields can be found at \url{https://www.infinitycloud.com/service/api/reports_v2_triggers.html}
}

