% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCaller.R
\name{PredictionCaller}
\alias{PredictionCaller}
\title{PredictionCaller}
\usage{
PredictionCaller(NormalizeList, model)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Performs the prediction and computes probability values. It also
}
