% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCaller.R
\name{PredictionCaller}
\alias{PredictionCaller}
\title{PredictionCaller}
\usage{
PredictionCaller(NormalizedData, model)
}
\arguments{
\item{NormalizedData}{A list of 3 vectors: normalized k-mer counts, genome length and contents of undefined bases.Produced by the´Kcounter´ function}

\item{model}{A random forest classification model. Must be FULL_HA or HA1. See \link{FULL_HA} and \link{HA1}.}
}
\value{
Data.frame with the classification results and quality checks.
The output has the following properties:
\itemize{
\item Each line corresponds to one sequence.
\item \code{Label} is the name of the sequence.
\item \code{Clade} is the corresponding prediction.
\item \code{Probability} is the proportions of trees that agreed with the Clade result. values between 0 to 1.
\item \code{Probability_QC} a logical value. If \code{TRUE} the sequence passed the quality filter for probability.
\item \code{Length} Sequence length.
\item \code{Length_QC} a logical value. If \code{TRUE} the sequence passed the quality filter for length.
\item \code{N} proportions of undefined bases in the sequence. The lower the better.
\item \code{N_QC} a logical value. If \code{TRUE} the sequence passed the quality filter for undefined bases.
}
}
\description{
Performs the prediction and computes probability values.
}
\examples{



file_path<-system.file("extdata","test_dataset.fasta",package="infinity")

sequence<-ape::read.FASTA(file_path,type = "DNA")

NormalizedData <- Kcounter(SequenceData=sequence,model=FULL_HA)

PredictedData <- PredictionCaller(NormalizedData=NormalizedData,model=FULL_HA)

}
