% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCaller.R
\name{Quality_filter}
\alias{Quality_filter}
\title{Quality_filter}
\usage{
Quality_filter(data)
}
\arguments{
\item{data}{data.frame obtained with \code{\link[=PredictionCaller]{PredictionCaller()}}}
}
\value{
data.frame
}
\description{
Change Clade definition to LowQuality in samples with either a FLAg from genome length or
proportion of N bases.
}
\examples{

file_path<-system.file("extdata","test_dataset.fasta",package="infinity")

sequence<-ape::read.FASTA(file_path,type = "DNA")

NormalizedData <- Kcounter(SequenceData=sequence,model=FULL_HA)

PredictedData <- PredictionCaller(NormalizedData=NormalizedData,model=FULL_HA)

PredictedData <- QualityControl(PredictedData,model=FULL_HA)

PredictedData <- Quality_filter(PredictedData)

}
