% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCaller.R
\name{QualityControl}
\alias{QualityControl}
\title{QualityControl}
\usage{
QualityControl(n_length, genome_length, probability)
}
\arguments{
\item{probability}{}
}
\value{

}
\description{
QualityControl
}
\examples{

}
