% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCaller.R
\name{QualityControl}
\alias{QualityControl}
\title{QualityControl}
\usage{
QualityControl(n_length, genome_length, probability, model)
}
\arguments{
\item{n_length}{numeric numeric}

\item{genome_length}{numeric numeric}

\item{probability}{numeric numeric}

\item{model}{numeric numeric}
}
\value{
A list with three logical vectors. In each case TRUE means pass.
}
\description{
QualityControl
}
