% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionCaller.R
\name{Stringent_filter}
\alias{Stringent_filter}
\title{Stringent_filter}
\usage{
Stringent_filter(data, QC_value = 0.2)
}
\arguments{
\item{data}{data.frame obtained with \code{\link[=PredictionCaller]{PredictionCaller()}}}

\item{QC_value}{numeric value from 0 to 1. (default = 0.6)}
}
\value{
data.frame
}
\description{
Change Clade definition to unknown in samples with a probability score
below a threshold
}
\examples{

file_path<-system.file("extdata","test_dataset.fasta",package="infinity")

sequence<-ape::read.FASTA(file_path,type = "DNA")

NormalizedData <- Kcounter(SequenceData=sequence,model=FULL_HA)

PredictedData <- PredictionCaller(NormalizedData=NormalizedData,model=FULL_HA)

PredictedData <- QualityControl(PredictedData,model=FULL_HA)

PredictedData <- Stringent_filter(PredictedData)

}
